Change FORCE_DL and NOFORCE
The FORCE_DL and NOFORCE enum is supplied to build_dep_graph for dependency
resolution purposes. When set to NOFORCE, PKGBUILDs for up to date AUR packages
will not be downloaded. When set to FORCE_DL, PKGBUILDS will be downloaded even
for up to date AUR packages.

With this patch, NOFORCE is renamed to RESOLVE_IMMEDIATE and FORCE_DL is renamed
to RESOLVE_THOROUGH. This is to prepare for an upcoming series which slightly
modifies the dependency resolution algorithm.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>